split class ScreenDimensionSupplier to be able to inject and Supplier of Screen
